Enh:Use functions and same terms in all files
